requirements-cloud.txt
jupyter==1.0.0
matplotlib==3.5.2
numpy==1.22.3
pandas==1.4.2
quinine==0.3.0
scikit-learn==1.0.2
seaborn==0.11.2
tqdm==4.64.0
transformers==4.17.0
wandb==0.12.11
xgboost==1.6.1
protobuf==3.20.1
torch==1.11.0
google-cloud-storage==2.7.0
google-cloud-aiplatform==1.25.0RUN pip install torch==1.11.0+cpu torchvision==0.12.0+cpu torchaudio==0.11.0 --extra-index-url https://download.pytorch.org/whl/cpu